% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpca.R
\name{xpca}
\alias{xpca}
\title{Cross-Covariance Principal Components Analysis}
\usage{
xpca(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\value{
list
}
\description{
Cross-Covariance Principal Components Analysis
}
\examples{
# Generate some example data
set.seed(123)
x <- matrix(rnorm(100 * 3), nrow = 100, ncol = 3)
y <- matrix(rnorm(100 * 5), nrow = 100, ncol = 5)
xpca(x, y)
}
